Adding basic seeding logic for users and user roles
This commit adds two new dependencies to the project: pino and faker-js/faker. It also creates a Seeder class that will be used to generate fake data for testing purposes. Two seeders are created as well: one for User records and another for UserRole records. Finally, a logger utility is added using the pino library.